--- a/digits_test.go
+++ b/digits_test.go
@@ -361,22 +361,22 @@ var testCases = []*testCase{
 			tail:       ")",
 		},
 	},
-	//{ // TODO BUG https://github.com/joshuanario/digits/issues/7
-	//	stimulus: stimulus{
-	//		precision:           digits.Tenth,
-	//		value:               "80800000.99090909090",
-	//		groupSeparator:      ',',
-	//		fractionalPrecision: digits.PreserveUpToHundredth,
-	//	},
-	//	expectation: expectation{
-	//		sigFigs:    "80800000.9",
-	//		nonSigFigs: "9",
-	//		strOut:     "80,800,000.99",
-	//		head:       "",
-	//		core:       "80,800,000.9",
-	//		tail:       "9",
-	//	},
-	//},
+	{
+		stimulus: stimulus{
+			precision:           digits.Tenth,
+			value:               "80800000.99090909090",
+			groupSeparator:      ',',
+			fractionalPrecision: digits.PreserveUpToHundredth,
+		},
+		expectation: expectation{
+			sigFigs:    "80800000.9",
+			nonSigFigs: "9",
+			strOut:     "80,800,000.99",
+			head:       "",
+			core:       "80,800,000.9",
+			tail:       "9",
+		},
+	},
 }
 
 func process(stimulus stimulus) *digits.Expression {