Closes #14 and closes #13 

Before:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkSigFigs$ github.com/joshuanario/digits

goos: darwin
goarch: arm64
pkg: github.com/joshuanario/digits
BenchmarkSigFigs-10    	   22150	     52840 ns/op	   31168 B/op	    1267 allocs/op
PASS
ok  	github.com/joshuanario/digits	1.833s
```

After:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkSigFigs$ github.com/joshuanario/digits

goos: darwin
goarch: arm64
pkg: github.com/joshuanario/digits
BenchmarkSigFigs-10    	   18769	     62229 ns/op	   37632 B/op	    1499 allocs/op
PASS
ok  	github.com/joshuanario/digits	1.929s
```

Edit: Close #7